% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Squiggle-methods.R
\docType{methods}
\name{meta,Squiggle-method}
\alias{meta,Squiggle-method}
\title{Extract signal metadata}
\usage{
\S4method{meta}{Squiggle}(object)
}
\arguments{
\item{object}{the Squiggle object}
}
\value{
list containing various metadata for transforming the signal to pA scale
}
\description{
Extract signal metadata
}
