% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Squiggle-methods.R
\name{plot_squiggle}
\alias{plot_squiggle}
\title{Plot nanopore signal}
\usage{
plot_squiggle(object, time_span = c(0, 0.5))
}
\arguments{
\item{object}{the Squiggle object to plot}

\item{time_span}{the time span in seconds to plot (default: c(0, 0.5))}
}
\description{
Plot nanopore signal
}
