% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PorexploreR-package.R
\docType{package}
\name{PorexploreR-package}
\alias{PorexploreR-package}
\alias{PorexploreR}
\title{A framework for benchmarking combinations of methods in multi-stage pipelines}
\description{
A framework for benchmarking combinations of methods in multi-stage pipelines
}
\seealso{
The core function in this package is \code{\link{Squiggle}}
}
\author{
Shian Su <\url{https://www.github.com/shians}>
}
