% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Squiggle-methods.R
\docType{methods}
\name{raw_signal,Squiggle-method}
\alias{raw_signal,Squiggle-method}
\title{Extract raw signal}
\usage{
\S4method{raw_signal}{Squiggle}(object)
}
\arguments{
\item{object}{the Squiggle object}
}
\value{
numeric vector of raw signal values
}
\description{
Extract raw signal
}
