% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Squiggle-methods.R
\docType{methods}
\name{signal,Squiggle-method}
\alias{signal,Squiggle-method}
\title{Extract signal}
\usage{
\S4method{signal}{Squiggle}(object)
}
\arguments{
\item{object}{the Squiggle object}
}
\value{
numeric vector of signal values transformed to pA scale
}
\description{
Extract signal
}
