% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Squiggle.R, R/Squiggle-methods.R
\docType{class}
\name{Squiggle-class}
\alias{Squiggle-class}
\alias{raw_signal}
\alias{signal}
\alias{meta}
\title{Squiggle}
\usage{
raw_signal(object)

signal(object)

meta(object)
}
\description{
This class is the representation of a "squiggle" from nanopore sequencing.
}
\section{Slots}{

\describe{
\item{\code{raw_signal}}{integer. Represents the raw signal values.}

\item{\code{signal}}{numeric (optional). Represents the transformed signal values.}

\item{\code{range}}{numeric. Represents the range of values in signal.}

\item{\code{digitisation}}{integer. Represents the digitisation range.}

\item{\code{offset}}{integer. Represents the offset for raw signal.}

\item{\code{sampling_rate}}{integer. Represents the raw signals captuerd per second.}

\item{\code{scaling}}{numeric. Represents the scaling factor when transform raw to original signal.}
}}

