% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Squiggle.R
\name{Squiggle}
\alias{Squiggle}
\title{Squiggle constructor}
\usage{
Squiggle(file_path)
}
\arguments{
\item{file_path}{path to a fast5 file}
}
\value{
Squiggle object
}
\description{
Construct a Squiggle object from a path to a fast5 file
}
