% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Squiggle-methods.R
\docType{methods}
\name{plot_squiggle}
\alias{plot_squiggle}
\alias{plot_squiggle,Squiggle,numeric-method}
\alias{plot_squiggle,Squiggle,missing-method}
\title{Plot nanopore signal}
\usage{
plot_squiggle(object, time_span = c(0, 0.5))

\S4method{plot_squiggle}{Squiggle,numeric}(object, time_span = c(0, 0.5))

\S4method{plot_squiggle}{Squiggle,missing}(object, time_span = c(0, 0.5))
}
\arguments{
\item{object}{the Squiggle object to plot}

\item{time_span}{the time span in seconds to plot (default: c(0, 0.5))}
}
\description{
Plot nanopore signal
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Squiggle,time_span = numeric}: 

\item \code{object = Squiggle,time_span = missing}: 
}}

